Fixme: adds page create boards with input mentioned
