+++ board/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from board.models import Board
+from board.models import Board, Mentioned
 
 
 admin.site.register(Board)
+admin.site.register(Mentioned)
