--- mentioned/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from mentioned.models import Mentioned
-
-
-admin.site.register(Mentioned)
