--- mentioned/urls.py
@@ -1,13 +1,0 @@
-from django.urls import path
-
-from board import views
-
-
-app_name = 'mentioned'
-urlpatterns = [
-  # path(
-  #   '',
-  #   views.FoodList.as_view(),
-  #   name='list'
-  # ),
-]
